+++ packages/knowledge/src/index.ts
@@ -5,3 +5,4 @@ export * from "./rules.js";
 export * from "./conformance.js";
 export * from "./rule-conformance.js";
 export * from "./drift.js";
+export * from "./discovery.js";